feat: remimplement txnData (#694)
* feat: remimplement txnData

* fix: debug

* fix(txnData): partly fix cumulativeGasUsed

* fix: cumulative gas in a block is sum of gas used - refund in each tx

* test(rlpAddr): too much gas

* feat(txndata): rename package from txn_data to txndata

* fix: changes in constraints

* fix: faulthy package import

---------

+++ tracer/arithmetization/src/main/java/net/consensys/linea/zktracer/module/euc/EucOperation.java
@@ -53,7 +53,7 @@ public class EucOperation extends ModuleOperation {
 
   public Bytes ceiling() {
     return !remainder.isZero() && !dividend.isZero()
-        ? Bytes.minimalBytes(quotient.toInt() + 1)
+        ? Bytes.minimalBytes(quotient.toLong() + 1)
         : quotient;
   }
